Cleanup some code, restructuring things
