+++ src/lua_engine/lua_stdlib/runtime_types.rs
@@ -74,97 +74,6 @@ impl Clone for SharedChannel {
     }
 }
 
-impl UserData for HalfEdgeMesh {
-    fn add_methods<'lua, M: mlua::UserDataMethods<'lua, Self>>(methods: &mut M) {
-        methods.add_method(
-            "get_channel",
-            |_lua, this, (kty, vty, name): (ChannelKeyType, ChannelValueType, String)| {
-                Ok(SharedChannel(
-                    this.channels
-                        .channel_rc_dyn(kty, vty, &name)
-                        .map_lua_err()?,
-                ))
-            },
-        );
-        methods.add_method(
-            "get_channel_2",
-            |lua, this, (kty, vty, name): (ChannelKeyType, ChannelValueType, String)| {
-                profiling::scope!("get_channel2");
-                use slotmap::Key;
-                let conn = this.read_connectivity();
-                let keys: Box<dyn Iterator<Item = u64>> = match kty {
-                    ChannelKeyType::VertexId => {
-                        Box::new(conn.iter_vertices().map(|(v_id, _)| v_id.data().as_ffi()))
-                    }
-                    ChannelKeyType::FaceId => {
-                        Box::new(conn.iter_faces().map(|(f_id, _)| f_id.data().as_ffi()))
-                    }
-                    ChannelKeyType::HalfEdgeId => {
-                        Box::new(conn.iter_halfedges().map(|(h_id, _)| h_id.data().as_ffi()))
-                    }
-                };
-                Ok(this
-                    .channels
-                    .dyn_read_channel_by_name(kty, vty, &name)
-                    .map_lua_err()?
-                    .to_table(keys, lua))
-            },
-        );
-        methods.add_method("set_channel", |lua, this, (kty, vty, name, table)| {
-            profiling::scope!("set_channel");
-            use slotmap::Key;
-            let name: String = name;
-            let conn = this.read_connectivity();
-            let keys: Box<dyn Iterator<Item = u64>> = match kty {
-                ChannelKeyType::VertexId => {
-                    Box::new(conn.iter_vertices().map(|(v_id, _)| v_id.data().as_ffi()))
-                }
-                ChannelKeyType::FaceId => {
-                    Box::new(conn.iter_faces().map(|(f_id, _)| f_id.data().as_ffi()))
-                }
-                ChannelKeyType::HalfEdgeId => {
-                    Box::new(conn.iter_halfedges().map(|(h_id, _)| h_id.data().as_ffi()))
-                }
-            };
-            this.channels
-                .dyn_write_channel_by_name(kty, vty, &name)
-                .map_lua_err()?
-                .set_from_table(keys, lua, table)
-                .map_lua_err()
-        });
-        methods.add_method_mut(
-            "ensure_channel",
-            |_lua, this, (kty, vty, name): (ChannelKeyType, ChannelValueType, String)| {
-                let _ = this.channels.ensure_channel_dyn(kty, vty, &name);
-                Ok(SharedChannel(
-                    // TODO: This needlesly recomputes the name->ch_id correspondence
-                    this.channels
-                        .channel_rc_dyn(kty, vty, &name)
-                        .map_lua_err()?,
-                ))
-            },
-        );
-        methods.add_method_mut("iter_vertices", |lua, this, ()| {
-            let vertices: Vec<VertexId> = this
-                .read_connectivity()
-                .iter_vertices()
-                .map(|(id, _)| id)
-                .collect();
-            let mut i = 0;
-            lua.create_function_mut(move |lua, ()| {
-                let val = if i < vertices.len() {
-                    vertices[i].to_lua(lua)?
-                } else {
-                    mlua::Value::Nil
-                };
-                i += 1;
-                Ok(val)
-            })
-        });
-        methods.add_method("clone", |_lua, this, ()| Ok(this.clone()));
-    }
-}
-
 #[test]
 pub fn minibench() {
     let lua = Lua::new();
